+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Added
+
+- Logging to a file
+- Ability for user to request an upload of the logs file
+
 ### Updated
 
 - Added basic spam reduction when logging so the same message won't get logged multiple times in a row